requirements.txt
annotated-types==0.7.0
anyio==4.8.0
bcrypt==4.2.1
boto3==1.36.21
botocore==1.36.21
certifi==2025.1.31
click==8.1.8
distro==1.9.0
dnspython==2.7.0
ecdsa==0.19.0
exceptiongroup==1.2.2
fastapi==0.115.8
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
iniconfig==2.0.0
jiter==0.8.2
jmespath==1.0.1
jose==1.0.0
mangum==0.19.0
motor==3.7.0
openai==1.63.0
packaging==24.2
passlib==1.7.4
pluggy==1.5.0
pyasn1==0.6.1
pydantic==2.10.6
pydantic_core==2.27.2
pymongo==4.11.1
pytest==8.3.4
pytest-asyncio==0.25.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.20
rsa==4.9
s3transfer==0.11.2
six==1.17.0
sniffio==1.3.1
starlette==0.45.3
tomli==2.2.1
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==1.26.0
uvicorn==0.34.0
